% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_cppcor.R
\name{GUI_cppcor}
\alias{GUI_cppcor}
\title{Graphical User Interface for PCCor function}
\usage{
GUI_cppcor(gui = TRUE)
}
\arguments{
\item{gui}{Logical argument, \code{TRUE} or \code{FALSE}. The default is \code{FALSE}}
}
\value{
\code{GUI_cppcor} return a GUI for the confusion matriz and parameters of classification analysis
}
\description{
\code{GUI_cppcor} A Graphical User Interface (GUI) for function that returns the confusion matrix and parameters of classification analysis
}
\examples{
# Loading package
library(cppcor)
if (interactive()) {
  GUI_cppcor()
}

}
